requirements.txt
langchain
llama_index
openai
protobuf
PyPDF2
python-dotenv
python_docx
Requests
streamlit
streamlit_chat
